+++ src/main/java/blusunrize/immersiveengineering/common/register/IEBlocks.java
@@ -165,7 +165,8 @@ public final class IEBlocks
 		public static final BlockEntry<IEBaseBlock> ALLOYBRICK = BlockEntry.simple("alloybrick", STONE_DECO_GBRICK_PROPS);
 		public static final BlockEntry<IEBaseBlock> SLAG_BRICK = BlockEntry.simple("slag_brick", STONE_DECO_GBRICK_PROPS);
 		public static final BlockEntry<IEBaseBlock> CLINKER_BRICK = BlockEntry.simple("clinker_brick", STONE_DECO_BRICK_PROPS);
-		public static final BlockEntry<IEBaseBlock> CLINKER_BRICK_SILL = BlockEntry.simple("clinker_brick_sill", STONE_DECO_BRICK_PROPS);
+		public static final BlockEntry<VerticalFacingBlock> CLINKER_BRICK_SILL = new BlockEntry<>(
+				"clinker_brick_sill", STONE_DECO_BRICK_PROPS, VerticalFacingBlock::new);
 		public static final BlockEntry<HorizontalFacingBlock> CLINKER_BRICK_QUOIN = new BlockEntry<>(
 				"clinker_brick_quoin", STONE_DECO_BRICK_PROPS, HorizontalFacingBlock::new);
 		public static final BlockEntry<IEBaseBlock> COKE = BlockEntry.simple(